Fixed most known issues
Specifically fixed ftTeam lineup being a reference to teamLineup such that whne teamLineup changed ftTeamLinup would change as well. Now it is a copy. Also noticed that rebounds not associated with a player might have the wrong team ID. Created a check that uses one of the next events to determine who the rebound should have counted for (for example team0 shot, unknown rebound, team1 shot implies the rebound was by team1).