Dealing with tech fts
How to consistently deal with tech fts? Consider sequence of made shot - tech foul - tech ft - made shot; the first team makes the first shot and shoots the tech, but if time elapses from shot to ft then the tech ft counts as a change in possession. However, currently the possession flips on first made shot, so during tech ft the 'current possession' is wrong. If we treat the tech as establishing possession then on previous made shot it looks like the team retains the ball.