Fixed end/start of period time and annoying exception
Fixed end of a period and start of a period being the same 'real time' by adding a small multiplier to the period in the chunk_pbp method.

Hacky fix to the sequence rebound-ft-sub-missed ft- time elapses -def rebound. NBA considers that sub having a possession due to the missed ft + def rebound, but because we check poss changes before processing subs, the subs never get the possession. You might think this would be fixed because of the rebound after the missed ft, but because the rebound before happens at the same time as the ft the 'next possession' already assigns it to the defensive team, so the code acts like the rebound on hte ft is offensive (and doesn't give out possesssions). Fix is by specifically checking for last missed fts in the sub method