Fixed some minor bugs
Noticed in Minnesota game that end of 4Q there was a shot clock turnover not assigned to a particular player. Now if 'Person1' is not a real player, code trusts 'Team_ID'.

Also noticed on two separate occassions (Collins and KAT) where they sub in on fts where the 1st ft is missed and it counts as their possession (I think because the playbyplay includes a rebound). However this is crazy and I will not do that